+++ test/test_computations/test_pureparallel.py
@@ -5,6 +5,7 @@ from helpers import *
 from reikna.pureparallel import PureParallel
 from reikna.core import Parameter, Annotation, Type
 import reikna.cluda.dtypes as dtypes
+from reikna.transformations import scale_param
 
 
 def test_guiding_input(thr):
@@ -90,3 +91,57 @@ def test_guiding_shape(thr):
     res_ref = numpy.vstack([a[0] + a[1], a[0] - a[1]])
 
     assert diff_is_negligible(res_dev.get(), res_ref)
+
+
+def test_trf_with_guiding_input(thr):
+    """
+    Test the creation of ``PureParallel`` out of a transformation,
+    with an input parameter as a guiding array.
+    """
+
+    N = 1000
+    coeff = 3
+    dtype = numpy.float32
+
+    arr_t = Type(dtype, shape=N)
+    trf = scale_param(arr_t, dtype)
+    p = PureParallel.from_trf(trf, trf.input)
+
+    # The new PureParallel has to preserve the parameter list of the original transformation.
+    assert list(p.signature.parameters.values()) == list(trf.signature.parameters.values())
+
+    a = get_test_array_like(p.parameter.input)
+    a_dev = thr.to_device(a)
+    res_dev = thr.empty_like(p.parameter.output)
+
+    pc = p.compile(thr)
+    pc(res_dev, a_dev, coeff)
+
+    assert diff_is_negligible(res_dev.get(), a * 3)
+
+
+def test_trf_with_guiding_output(thr):
+    """
+    Test the creation of ``PureParallel`` out of a transformation,
+    with an output parameter as a guiding array.
+    """
+
+    N = 1000
+    coeff = 3
+    dtype = numpy.float32
+
+    arr_t = Type(dtype, shape=N)
+    trf = scale_param(arr_t, dtype)
+    p = PureParallel.from_trf(trf, trf.output)
+
+    # The new PureParallel has to preserve the parameter list of the original transformation.
+    assert list(p.signature.parameters.values()) == list(trf.signature.parameters.values())
+
+    a = get_test_array_like(p.parameter.input)
+    a_dev = thr.to_device(a)
+    res_dev = thr.empty_like(p.parameter.output)
+
+    pc = p.compile(thr)
+    pc(res_dev, a_dev, coeff)
+
+    assert diff_is_negligible(res_dev.get(), a * 3)