+++ doc/source/history.rst
@@ -8,6 +8,8 @@ Release history
 
 * ADDED: the first argument to the ``Transformation`` or ``PureParallel`` snippet is now a ``reikna.core.Indices`` object instead of a list.
 
+* ADDED: classmethod ``PureParallel.from_trf()``, which allows one to create a pure parallel computation out of a transformation.
+
 * FIXED: improved ``Computation.compile()`` performance for complicated computations by precreating transformation templates.